Make panel-store a ports-and-adapters unit test
